rename InjectPropertyRector to AnnotatedPropertyInjectToConstructorInjectionRector
