+++ packages/Nette/tests/Rector/Application/AnnotatedPropertyInjectToConstructorInjectionRector/Correct/correct.php.inc
