+++ packages/Nette/tests/Rector/Application/AnnotatedPropertyInjectToConstructorInjectionRector/Wrong/wrong.php.inc
