+++ packages/Nette/tests/Rector/Application/AnnotatedPropertyInjectToConstructorInjectionRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\Nette\Rector\Application\AnnotatedPropertyInjectToConstructorInjectionRector: ~
