--- packages/Nette/tests/Rector/Application/InjectPropertyRector/config.yml
@@ -1,2 +1,0 @@
-services:
-    Rector\Nette\Rector\Application\InjectPropertyRector: ~
